requirements.txt
scipy
jupyter
Pyarrow
pandas
numpy
matplotlib
openpyxl
xlsxwriter
scipy
scikit-learn
seaborn
factor_analyzer
mapclassify
python-docx
js2py
PyMuPDF
retrying
parsel
jieba
xgboost
catboost
lightgbm
